# api/openapi.yaml
openapi: 3.1.0
info:
  title: ForgeERP API
  version: 0.1.0
  description: >
    ForgeERP modernized ERP/CRM API (Dolibarr functional successor).
    This document is authoritative for REST surface; phase routers must register
    every path declared here.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/finance/entries:
    post:
      operationId: postEntry
      summary: Validate, chain-link and post an entry (requires finance.entry.write)
      responses:
        '201': { description: Posted }
        '422': { description: Unbalanced or locked year }
...
  /api/v1/hr/leaves:
    get:
      operationId: listLeaves
      summary: Page leave requests, optionally by user (requires hr.leave.read)
      responses:
        '200': { description: List }
    post:
      operationId: createLeave
      summary: File a draft leave request (requires hr.leave.write)
      responses:
        '201': { description: Created }
        '422': { description: Validation failed }
  /api/v1/hr/leaves/{id}/status:
    post:
      operationId: setLeaveStatus
      summary: State-machine transition for a leave (requires hr.leave.validate)
      responses:
        '200': { description: Updated }
        '422': { description: Illegal transition }
  /api/v1/hr/expenses:
    get:
      operationId: listExpenses
      summary: Page expense reports, optionally by user (requires hr.expense.read)
      responses:
        '200': { description: List }
    post:
      operationId: createExpense
      summary: Open a draft expense report (requires hr.expense.write)
      responses:
        '201': { description: Created }
        '409': { description: Duplicate ref }
        '422': { description: Validation failed }
  /api/v1/hr/expenses/{id}/lines:
    post:
      operationId: addExpenseLine
      summary: Append a line to a draft report (requires hr.expense.write)
      responses:
        '201': { description: Created }
        '422': { description: Report not editable }
  /api/v1/hr/expenses/{id}/pay:
    post:
      operationId: payExpense
      summary: Pay an approved report posting a balanced entry (requires hr.expense.pay)
      responses:
        '200': { description: Paid }
        '422': { description: Not approved or bad accounts }
        '502': { description: Ledger failed, reverted }
        '503': { description: Finance not wired }
  /api/v1/hr/salaries:
    get:
      operationId: listSalaries
      summary: List salary records, optionally by user (requires hr.salary.read)
      responses:
        '200': { description: List }
    post:
      operationId: createSalary
      summary: Record a draft salary (requires hr.salary.write)
      responses:
        '201': { description: Created }
        '422': { description: Unbalanced or invalid }
...